backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis[asyncio]==5.0.8
httpx==0.27.2
yfinance==0.2.43
pydantic==2.9.2
python-multipart==0.0.9
python-dotenv==1.0.1
openai==1.51.0
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
prophet==1.1.5
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
APScheduler==3.10.4
